a fuction created

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -2,7 +2,7 @@
 -- returns 0 if the second number is equal to 0.
 DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS INT
+RETURNS FLOAT
 BEGIN
     IF b = 0 THEN
         RETURN 0;
